clean sdk
